pyproject.toml
[tool.poetry]
name = "PocketLeague"
version = "0.1.0"
description = ""
authors = ["marcelm9 <marcel.menzel09@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "PocketLeague", from = "main" }]

[tool.poetry.dependencies]
python = "^3.11"
pygame-ce = "^2.4.0"
pygamextras = "^1.1.4"
pymunk = "^6.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
